requirements.txt

asgiref==3.6.0
Django==2.2.19
django-cors-headers==3.14.0
djangorestframework==3.12.4
django-filter==23.2
cffi==1.15.1
importlib-metadata==4.2.0
mccabe==0.7.0
pip==23.1.2
pillow==9.5.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2023.3
psycopg2-binary==2.8.6
pycairo==1.23.0
python-dotenv==0.21.1
setuptools==47.1.0
sqlparse==0.4.4
reportlab==4.0.0
typing_extensions==4.5.0
zipp== 3.15.0